Tests: Function start using pytest
Rearranged the test class to be usable directly by calling pytest.
